+++ src/main/java/group/msg/at/cloud/common/observability/rest/trace/RestTraceConstants.java
@@ -1,4 +1,4 @@
-package group.msg.at.cloud.common.observability.spring.rest.trace;
+package group.msg.at.cloud.common.observability.rest.trace;
 
 public interface RestTraceConstants {