+++ changelog.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-02-28
+### Added
+### Changed
+- refactored package names by dropping the `spring` subpackage from the hierarchy
+- added support for logging context information via a new LoggingMdcFilter plus auto-configuration.
+
 ## [0.1.0] - 2022-02-11
 ### Added
 ### Changed